+++ src/generators/content/GeneratorExpression.ts
@@ -1,19 +1,23 @@
 import { FormatSpec } from "../../utils/FormatSpec";
 import { IGeneratorExpression } from "./IGeneratorExpression";
 
-const _expressionRegEx = /{(?!})(?:(\w+)|#((?:\-?\()*\d+d\d+(?: [+\-*/] (?:\-?\()*(?:\d+d\d+|\-?\d+)\)*)*))?(?:(?<!{)\[((?:[^\n{}]?[<>=^])?[+\- ]?#?0?\d*[_,]?(?:\.\d+)?[bcdeEfFgGnosxX%]?)\])?(?::(\w+))?}/;
+const _expressionRegEx = /{(?!})(?:(\w+)(?:<(\d+)(?:\|([^>]+)?)>)?|#((?:\-?\()*\d+d\d+(?: [+\-*/] (?:\-?\()*(?:\d+d\d+|\-?\d+)\)*)*))?(?:(?<!{)\[((?:[^\n{}]?[<>=^])?[+\- ]?#?0?\d*[_,]?(?:\.\d+)?[bcdeEfFgGnosxX%]?)\])?(?::(\w+))?}/;
 
 enum TemplateMatch {
     Whole = 0,
     GeneratorName = 1,
-    DiceRoll = 2,
-    FormatSpec = 3,
-    VariableName = 4
+    GeneratorRepeat = 2,
+    RepeatSeparator = 3,
+    DiceRoll = 4,
+    FormatSpec = 5,
+    VariableName = 6
 }
 
 export class GeneratorExpression implements IGeneratorExpression {
     private _wholeMatch: string;
     private _generatorName?: string;
+    private _generatorRepeat?: number;
+    private _repeatSeparator?: string;
     private _variableName?: string;
     private _diceRoll?: string;
     private _formatSpec?: FormatSpec;
@@ -21,6 +25,11 @@ export class GeneratorExpression implements IGeneratorExpression {
     private constructor(matches: RegExpMatchArray) {
         this._wholeMatch = matches[TemplateMatch.Whole];
         this._generatorName = matches[TemplateMatch.GeneratorName];
+        let repeat = parseInt(matches[TemplateMatch.GeneratorRepeat]);
+        if (!isNaN(repeat)) {
+            this._generatorRepeat = repeat;
+        }
+        this._repeatSeparator = matches[TemplateMatch.RepeatSeparator];
         this._variableName = matches[TemplateMatch.VariableName];
         this._diceRoll = matches[TemplateMatch.DiceRoll];
         this._formatSpec = FormatSpec.getFormatSpec(matches[TemplateMatch.FormatSpec]);
@@ -40,6 +49,14 @@ export class GeneratorExpression implements IGeneratorExpression {
 
     public get generatorName(): string | undefined {
         return this._generatorName;
+    }
+
+    public get generatorRepeat(): number | undefined {
+        return this._generatorRepeat;
+    }
+
+    public get repeatSeparator(): string | undefined {
+        return this._repeatSeparator;
     }
 
     public get wholeMatch(): string {