+++ src/generators/content/IGeneratorExpression.ts
@@ -2,8 +2,10 @@ import { FormatSpec } from "../../utils/FormatSpec";
 
 export interface IGeneratorExpression {
     wholeMatch: string;
-    variableName: string | undefined;
     generatorName: string | undefined;
+    generatorRepeat: number | undefined;
+    repeatSeparator: string | undefined;
     diceRoll: string | undefined;
     formatSpec: FormatSpec | undefined;
+    variableName: string | undefined;
 }